pyproject.toml
[project]
name = "voxel51-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.9.0",
    "datasets>=4.0.0",
    "fiftyone==1.7.0",
    "huggingface>=0.0.1",
    "lightning>=2.5.2",
    "open-clip-torch>=3.0.0",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "spyder-kernels==3.0.*",
    "timm>=1.0.19",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "transformers>=4.54.1",
    "ultralytics>=8.3.171",
]
